Refactored PreferencesTab file structure
